Clean up post-merge duplicate members in TrxReportEngine
After merging origin/main (which split TrxReportEngine into focused
partial files), the conflict resolution left duplicate definitions of
utility helpers. Reset TrxReportEngine.cs to main's slim version and
applied only the GenerateReportAsync signature change. Updated the
Results / Results.Definitions partials to use the new TrxTestResult DTO
instead of TestNodeUpdateMessage / TestNode.
