+++ src/Platform/Microsoft.Testing.Extensions.TrxReport/TrxReportEngine.Results.cs
@@ -1,17 +1,16 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Microsoft.Testing.Extensions.TrxReport.Abstractions.Streaming;
 using Microsoft.Testing.Extensions.TrxReport.Resources;
 using Microsoft.Testing.Platform;
-using Microsoft.Testing.Platform.Extensions.Messages;
 using Microsoft.Testing.Platform.Helpers;
-using Microsoft.Testing.Platform.Messages;
 
 namespace Microsoft.Testing.Extensions.TrxReport.Abstractions;
 
 internal sealed partial class TrxReportEngine
 {
-    private SummaryCounts AddResults(TestNodeUpdateMessage[] testNodeUpdateMessages, string testAppModule, XElement testRun, string runDeploymentRoot, XElement testDefinitions, XElement testEntries)
+    private SummaryCounts AddResults(IReadOnlyList<TrxTestResult> testResults, string testAppModule, XElement testRun, string runDeploymentRoot, XElement testDefinitions, XElement testEntries)
     {
         int passed = 0;
         int failed = 0;
@@ -25,22 +24,20 @@ internal sealed partial class TrxReportEngine
         // However, due to backcompat concerns, we will disallow this only for frameworks that start using TrxTestDefinitionName property.
         var uniqueTestDefinitionTestIds = new Dictionary<string, (string TestDefinitionName, bool IsExplicitlyProvided)>();
 
-        foreach (TestNodeUpdateMessage nodeMessage in testNodeUpdateMessages)
+        foreach (TrxTestResult testResult in testResults)
         {
-            TestNode testNode = nodeMessage.TestNode;
-
             // If already a guid (it's the case for at least MSTest), use that guid directly.
             // Otherwise, convert the string to a guid.
-            if (!Guid.TryParse(testNode.Uid.Value, out Guid guid))
+            if (!Guid.TryParse(testResult.Uid, out Guid guid))
             {
-                guid = GuidFromString(testNode.Uid.Value);
+                guid = GuidFromString(testResult.Uid);
             }
 
             // NOTE: In VSTest, MSTestDiscoverer.TmiTestId property is preferred if present.
             string id = guid.ToString();
-            string testResultDisplayName = RemoveInvalidXmlChar(testNode.DisplayName)!;
-            (string testDefinitionName, bool isExplicitlyProvided) = testNode.Properties.SingleOrDefault<TrxTestDefinitionName>() is { } trxTestDefinitionName
-                ? (RemoveInvalidXmlChar(trxTestDefinitionName.TestDefinitionName), true)
+            string testResultDisplayName = RemoveInvalidXmlChar(testResult.DisplayName)!;
+            (string testDefinitionName, bool isExplicitlyProvided) = testResult.TrxTestDefinitionName is { } explicitName
+                ? (RemoveInvalidXmlChar(explicitName)!, true)
                 : (testResultDisplayName, false);
 
             string executionId = Guid.NewGuid().ToString();
@@ -53,18 +50,17 @@ internal sealed partial class TrxReportEngine
                 new XAttribute("testName", testResultDisplayName),
                 new XAttribute("computerName", _environment.MachineName));
 
-            TimingProperty? timing = testNode.Properties.SingleOrDefault<TimingProperty>();
-            string testDuration = timing?.GlobalTiming.Duration is { } duration
+            string testDuration = testResult.Duration is { } duration
                 ? duration.ToString("hh\\:mm\\:ss\\.fffffff", CultureInfo.InvariantCulture)
                 : "00:00:00";
             unitTestResult.SetAttributeValue("duration", testDuration);
 
             unitTestResult.SetAttributeValue(
                 "startTime",
-                timing?.GlobalTiming.StartTime.ToUniversalTime().ToString("O") ?? _clock.UtcNow.ToString("O"));
+                testResult.StartTime?.ToUniversalTime().ToString("O") ?? _clock.UtcNow.ToString("O"));
             unitTestResult.SetAttributeValue(
                 "endTime",
-                timing?.GlobalTiming.EndTime.ToUniversalTime().ToString("O") ?? _clock.UtcNow.ToString("O"));
+                testResult.EndTime?.ToUniversalTime().ToString("O") ?? _clock.UtcNow.ToString("O"));
 
             // In VSTest, other test types originate from adding TestProperty with
             // Id TestType (see Constants.TestTypePropertyIdentifier).
@@ -72,43 +68,30 @@ internal sealed partial class TrxReportEngine
             // In the context of MTP, we don't care.
             unitTestResult.SetAttributeValue("testType", UnitTestTypeGuid);
 
-            string currentTestOutcome = "Passed";
+            string currentTestOutcome = testResult.Outcome switch
+            {
+                TrxTestOutcome.Skipped => "NotExecuted",
+                TrxTestOutcome.Passed => "Passed",
+                TrxTestOutcome.Failed or TrxTestOutcome.Timeout => "Failed",
+                _ => throw ApplicationStateGuard.Unreachable(),
+            };
 
-            // In TrxReportGenerator.ConsumeAsync, we already filtered to only the nodes that contain TestNodeStateProperty.
-            // We also filtered out discovered and in-progress states.
-            // So the call to Single here should never fail, and should never be discovered or in-progress.
-            TestNodeStateProperty testState = testNode.Properties.Single<TestNodeStateProperty>();
-            if (testState is DiscoveredTestNodeStateProperty or InProgressTestNodeStateProperty)
+            switch (testResult.Outcome)
             {
-                throw ApplicationStateGuard.Unreachable();
-            }
-
-            if (testState is SkippedTestNodeStateProperty)
-            {
-                currentTestOutcome = "NotExecuted";
-                skipped++;
-            }
-            else if (testState is PassedTestNodeStateProperty)
-            {
-                passed++;
-            }
-            else if (Array.IndexOf(TestNodePropertiesCategories.WellKnownTestNodeTestRunOutcomeFailedProperties, testState.GetType()) >= 0)
-            {
-                currentTestOutcome = "Failed";
-
-                if (testState is TimeoutTestNodeStateProperty)
-                {
+                case TrxTestOutcome.Skipped:
+                    skipped++;
+                    break;
+                case TrxTestOutcome.Passed:
+                    passed++;
+                    break;
+                case TrxTestOutcome.Timeout:
                     timedout++;
-                }
-                else
-                {
+                    break;
+                case TrxTestOutcome.Failed:
                     failed++;
-                }
-            }
-            else
-            {
-                // Above conditions should have handled all state properties.
-                throw ApplicationStateGuard.Unreachable();
+                    break;
+                default:
+                    throw ApplicationStateGuard.Unreachable();
             }
 
             unitTestResult.SetAttributeValue("outcome", currentTestOutcome);
@@ -124,38 +107,40 @@ internal sealed partial class TrxReportEngine
             // i.e. https://github.com/dotnet/runtime/blob/main/src/libraries/System.Private.Xml/src/System/Xml/Core/XmlEncodedRawTextWriter.cs#L890
             var output = new XElement("Output");
 
-            TrxMessagesProperty? trxMessages = testNode.Properties.SingleOrDefault<TrxMessagesProperty>();
-            IEnumerable<string?>? nonErrorMessages = trxMessages?.Messages.Where(x => x is not StandardErrorTrxMessage and not DebugOrTraceTrxMessage).Select(x => x.Message);
-            if (nonErrorMessages?.Any() == true)
+            IReadOnlyList<TrxStreamMessage>? trxMessages = testResult.Messages;
+            if (trxMessages is not null)
             {
-                output.Add(new XElement("StdOut", RemoveInvalidXmlChar(string.Join(Environment.NewLine, nonErrorMessages))));
+                IEnumerable<string?> nonErrorMessages = trxMessages.Where(x => x.Kind == TrxStreamMessageKind.StandardOutput).Select(x => x.Message);
+                if (nonErrorMessages.Any())
+                {
+                    output.Add(new XElement("StdOut", RemoveInvalidXmlChar(string.Join(Environment.NewLine, nonErrorMessages))));
+                }
+
+                IEnumerable<string?> errorMessages = trxMessages.Where(x => x.Kind == TrxStreamMessageKind.StandardError).Select(x => x.Message);
+                if (errorMessages.Any())
+                {
+                    output.Add(new XElement("StdErr", RemoveInvalidXmlChar(string.Join(Environment.NewLine, errorMessages))));
+                }
+
+                IEnumerable<string?> debugOrTraceMessages = trxMessages.Where(x => x.Kind == TrxStreamMessageKind.DebugOrTrace).Select(x => x.Message);
+                if (debugOrTraceMessages.Any())
+                {
+                    output.Add(new XElement("DebugTrace", RemoveInvalidXmlChar(string.Join(Environment.NewLine, debugOrTraceMessages))));
+                }
             }
 
-            IEnumerable<string?>? errorMessages = trxMessages?.Messages.Where(x => x is StandardErrorTrxMessage).Select(x => x.Message);
-            if (errorMessages?.Any() == true)
-            {
-                output.Add(new XElement("StdErr", RemoveInvalidXmlChar(string.Join(Environment.NewLine, errorMessages))));
-            }
-
-            IEnumerable<string?>? debugOrTraceMessages = trxMessages?.Messages.Where(x => x is DebugOrTraceTrxMessage).Select(x => x.Message);
-            if (debugOrTraceMessages?.Any() == true)
-            {
-                output.Add(new XElement("DebugTrace", RemoveInvalidXmlChar(string.Join(Environment.NewLine, debugOrTraceMessages))));
-            }
-
-            TrxExceptionProperty? trxException = testNode.Properties.SingleOrDefault<TrxExceptionProperty>();
-            if (trxException?.Message is not null || trxException?.StackTrace is not null)
+            if (testResult.ExceptionMessage is not null || testResult.ExceptionStackTrace is not null)
             {
                 XElement errorInfoElement = new("ErrorInfo");
 
-                if (trxException.Message is not null)
+                if (testResult.ExceptionMessage is not null)
                 {
-                    errorInfoElement.Add(new XElement("Message", RemoveInvalidXmlChar(trxException.Message)));
+                    errorInfoElement.Add(new XElement("Message", RemoveInvalidXmlChar(testResult.ExceptionMessage)));
                 }
 
-                if (trxException.StackTrace is not null)
+                if (testResult.ExceptionStackTrace is not null)
                 {
-                    errorInfoElement.Add(new XElement("StackTrace", RemoveInvalidXmlChar(trxException.StackTrace)));
+                    errorInfoElement.Add(new XElement("StackTrace", RemoveInvalidXmlChar(testResult.ExceptionStackTrace)));
                 }
 
                 output.Add(errorInfoElement);
@@ -172,14 +157,17 @@ internal sealed partial class TrxReportEngine
             // 2. CollectorDataEntries
             // So far, we only have "ResultFiles".
             XElement? resultFiles = null;
-            foreach (FileArtifactProperty testFileArtifact in testNode.Properties.OfType<FileArtifactProperty>())
+            if (testResult.FileArtifacts is not null)
             {
-                resultFiles ??= new XElement("ResultFiles");
+                foreach (TrxTestFileArtifact testFileArtifact in testResult.FileArtifacts)
+                {
+                    resultFiles ??= new XElement("ResultFiles");
 
-                string href = CopyArtifactIntoTrxDirectoryAndReturnHrefValue(testFileArtifact.FileInfo, runDeploymentRoot, executionId);
-                resultFiles.Add(new XElement(
-                    "ResultFile",
-                    new XAttribute("path", href)));
+                    string href = CopyArtifactIntoTrxDirectoryAndReturnHrefValue(new FileInfo(testFileArtifact.FullPath), runDeploymentRoot, executionId);
+                    resultFiles.Add(new XElement(
+                        "ResultFile",
+                        new XAttribute("path", href)));
+                }
             }
 
             if (resultFiles is not null)
@@ -215,7 +203,7 @@ internal sealed partial class TrxReportEngine
             else
             {
                 uniqueTestDefinitionTestIds.Add(id, (testDefinitionName, isExplicitlyProvided));
-                XElement unitTest = CreateUnitTestElementForTestDefinition(testDefinitionName, testAppModule, id, testNode, executionId);
+                XElement unitTest = CreateUnitTestElementForTestDefinition(testDefinitionName, testAppModule, id, testResult, executionId);
 
                 var testMethod = new XElement(
                     "TestMethod",
@@ -223,7 +211,7 @@ internal sealed partial class TrxReportEngine
                     new XAttribute("adapterTypeName", $"executor://{_testFrameworkAdapter.Uid}/{_testFrameworkAdapter.Version}"));
 
                 // NOTE: className is required by TRX XSD.
-                (string className, string? testMethodName) = GetClassAndMethodName(testNode);
+                (string className, string? testMethodName) = GetClassAndMethodName(testResult);
                 testMethod.SetAttributeValue("className", className);
 
                 // NOTE: Historically, MTP used to always use testResultDisplayName here.
@@ -259,22 +247,22 @@ internal sealed partial class TrxReportEngine
         return new SummaryCounts(passed, failed, skipped, timedout);
     }
 
-    private (string ClassName, string? TestMethodName) GetClassAndMethodName(TestNode testNode)
+    private (string ClassName, string? TestMethodName) GetClassAndMethodName(TrxTestResult testResult)
     {
-        TestMethodIdentifierProperty? testMethodIdentifierProperty = testNode.Properties.SingleOrDefault<TestMethodIdentifierProperty>();
+        TrxTestMethodIdentifier? methodIdentifier = testResult.TestMethodIdentifier;
 
-        if (testNode.Properties.SingleOrDefault<TrxFullyQualifiedTypeNameProperty>()?.FullyQualifiedTypeName is { } className)
+        if (testResult.TrxFullyQualifiedTypeName is { } className)
         {
-            return (className, testMethodIdentifierProperty?.MethodName);
+            return (className, methodIdentifier?.MethodName);
         }
 
-        _ = testMethodIdentifierProperty ?? throw new InvalidOperationException(string.Format(CultureInfo.InvariantCulture, ExtensionResources.TrxReportFrameworkDoesNotSupportTrxReportCapability, _testFrameworkAdapter.DisplayName, _testFrameworkAdapter.Uid));
+        _ = methodIdentifier ?? throw new InvalidOperationException(string.Format(CultureInfo.InvariantCulture, ExtensionResources.TrxReportFrameworkDoesNotSupportTrxReportCapability, _testFrameworkAdapter.DisplayName, _testFrameworkAdapter.Uid));
 
-        string classNameFromIdentifierProperty = RoslynString.IsNullOrEmpty(testMethodIdentifierProperty.Namespace)
-            ? testMethodIdentifierProperty.TypeName
-            : $"{testMethodIdentifierProperty.Namespace}.{testMethodIdentifierProperty.TypeName}";
+        string classNameFromIdentifierProperty = RoslynString.IsNullOrEmpty(methodIdentifier.Namespace)
+            ? methodIdentifier.TypeName
+            : $"{methodIdentifier.Namespace}.{methodIdentifier.TypeName}";
 
         // TODO: Are we expected to append backtick and arity here for generic methods?
-        return (classNameFromIdentifierProperty, testMethodIdentifierProperty.MethodName);
+        return (classNameFromIdentifierProperty, methodIdentifier.MethodName);
     }
 }
